Change the types of the edgeTable and triTable
The range on values in these tables is much less than needs an int so
making them specific width types packs them smaller.

I measure a 5% decrease in the size of the .so file created at -O3.

+++ library/PolyVoxCore/include/PolyVoxCore/Impl/MarchingCubesTables.h
@@ -28,8 +28,8 @@ freely, subject to the following restrictions:
 
 namespace PolyVox
 {
-	extern const POLYVOX_API int edgeTable[256];
-	extern const POLYVOX_API int triTable[256][16];
+	extern const POLYVOX_API uint16_t edgeTable[256];
+	extern const POLYVOX_API int8_t triTable[256][16];
 }
 
 #endif
